plotly's api is still confusing but I figured out how they want to specify a layout object. wish they published @types.

+++ draw-timeseries-plot/built/logEntry.js
@@ -37,7 +37,7 @@ class LogEntry {
         let start = self.Text.indexOf(ScoreEntryMarker) + ScoreEntryMarker.length;
         let end = self.Text.indexOf(",", start);
         let score = self.Text.substring(start, end);
-        return parseInt(score);
+        return parseFloat(score);
     }
 }
 exports.LogEntry = LogEntry;
