+++ draw-timeseries-plot/src/logEntry.ts
@@ -55,7 +55,7 @@ class LogEntry {
        let start = self.Text.indexOf(ScoreEntryMarker) + ScoreEntryMarker.length;
        let end = self.Text.indexOf(",", start);
        let score = self.Text.substring(start, end);
-       return parseInt(score);
+       return parseFloat(score);
     }
 }
